Move test cases under typecast tests
