--- tests/jballerina-unit-test/src/test/java/org/ballerinalang/test/types/readonly/ImmutableArrayTypeTest.java
@@ -1,71 +1,0 @@
-/*
- *  Copyright (c) 2021, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
- *
- *  WSO2 Inc. licenses this file to you under the Apache License,
- *  Version 2.0 (the "License"); you may not use this file except
- *  in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing,
- *  software distributed under the License is distributed on an
- *  "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- *  KIND, either express or implied.  See the License for the
- *  specific language governing permissions and limitations
- *  under the License.
- */
-
-package org.ballerinalang.test.types.readonly;
-
-import org.ballerinalang.test.BCompileUtil;
-import org.ballerinalang.test.BRunUtil;
-import org.ballerinalang.test.CompileResult;
-import org.testng.annotations.AfterClass;
-import org.testng.annotations.BeforeClass;
-import org.testng.annotations.DataProvider;
-import org.testng.annotations.Test;
-
-/**
- * Tests for cast of readonly array.
- */
-public class ImmutableArrayTypeTest {
-
-    private CompileResult result;
-
-    @BeforeClass
-    public void setup() {
-        result = BCompileUtil.compile("test-src/types/readonly/test_immutable_array_type.bal");
-    }
-
-    @Test(dataProvider = "immutableArrayTypesTestFunctions")
-    public void testImmutableArrayTypes(String function) {
-        BRunUtil.invoke(result, function);
-    }
-
-    @DataProvider(name = "immutableArrayTypesTestFunctions")
-    public Object[][] immutableArrayTypesTestFunctions() {
-        return new Object[][]{
-                {"testCastOfReadonlyIntArrayToByteArray"},
-                {"testCastOfReadonlyIntArrayToByteArrayNegative"},
-                {"testCastOfReadonlyAnyToByteArray"},
-                {"testCastOfReadonlyArrayToUnion"},
-                {"testCastOfReadonlyUnionArrayToByteArray"},
-                {"testCastOfReadonlyRecord"},
-                {"testCastOfReadonlyStringArrayToStringConstantArray"},
-                {"testCastOfTwoDimensionalIntArrayToByteArray"}
-        };
-    }
-
-    @Test(expectedExceptions = org.ballerinalang.core.util.exceptions.BLangRuntimeException.class,
-            expectedExceptionsMessageRegExp = ".*error: \\{ballerina\\}TypeCastError \\{\"message\":\"incompatible " +
-                    "types: '\\(Foo & readonly\\)' cannot be cast to 'Bar'\"\\}.*")
-    public void testCastOfReadonlyRecordNegative() {
-        BRunUtil.invoke(result, "testCastOfReadonlyRecordNegative");
-    }
-
-    @AfterClass
-    public void tearDown() {
-        result = null;
-    }
-}
